separate create_normal_token.c
