+++ srcs/create_token_util.c
@@ -6,34 +6,34 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/13 15:10:32 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/16 15:58:25 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/18 15:52:32 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"../includes/minishell.h"
 
-bool	create_token_word(int *start, char *line, t_token **head)
-{
-	char	*word;
-	int		j;
-	int		len;
-	bool	err_flg;
+// bool	create_token_word(int *start, char *line, t_token **head)
+// {
+// 	char	*word;
+// 	int		j;
+// 	int		len;
+// 	bool	err_flg;
 
-	err_flg = false;
-	len = (int)ft_strlen(line);
-	j = *start + 1;
-	while (!is_split_char(line[j]) && !is_quart(line[j]) && j < len)
-		j++;
-	word = ft_substr(line, *start, j - *start);
-	if (!word)
-		return (true);
-	if (*start == 0 || is_split_char(line[*start - 1]))
-		err_flg = add_token(head, create_token_node(word));
-	else
-		err_flg = add_current_token(word, head);
-	*start = j;
-	return (err_flg);
-}
+// 	err_flg = false;
+// 	len = (int)ft_strlen(line);
+// 	j = *start + 1;
+// 	while (!is_split_char(line[j]) && !is_quart(line[j]) && j < len)
+// 		j++;
+// 	word = ft_substr(line, *start, j - *start);
+// 	if (!word)
+// 		return (true);
+// 	if (*start == 0 || is_split_char(line[*start - 1]))
+// 		err_flg = add_token(head, create_token_node(word));
+// 	else
+// 		err_flg = add_current_token(word, head);
+// 	*start = j;
+// 	return (err_flg);
+// }
 
 t_token	*create_token_node(char *word)
 {
@@ -79,4 +79,19 @@ bool	is_quart(char c)
 	if (c == '\'' || c == '\"')
 		return (true);
 	return (false);
+}
+
+bool	add_current_token(char *word, t_token **head)
+{
+	char	*temp;
+	bool	err_flg;
+
+	err_flg = false;
+	temp = ft_strjoin(get_last_token(head)->word, word);
+	free(get_last_token(head)->word);
+	free(word);
+	if (!temp)
+		return (true);
+	get_last_token(head)->word = temp;
+	return (err_flg);
 }
