+++ srcs/parser.c
@@ -6,25 +6,11 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/07 18:22:31 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/16 16:22:41 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/18 16:04:41 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
-
-typedef struct s_cmd
-{
-	char	*cmd; // いらない
-	char	**args; // tokenを突っ込む　ls, > , in ,
-	char	*cmd_abs_path;
-	int		status;
-} t_cmd;
-
-typedef struct s_cmd_list
-{
-	t_cmd	cmd;
-	t_cmd	*next;
-}	t_cmd_list;
 
 bool	parser(char *line)
 {
@@ -32,7 +18,7 @@ bool	parser(char *line)
 	bool	err_flg;
 
 	err_flg = false;
-	// lineが長すぎる場合はここでエラー返す
+	// TODO: line too long
 
 	head = NULL;
 	if (validation_quart(line))
@@ -40,10 +26,6 @@ bool	parser(char *line)
 		syntax_error(&head);
 		return (true);
 	}
-// ============ OK =========================
-
-//
-	// create_token_list(line, &head);
 	head = create_token_list(line);
 	if (!head)
 		return (true);
@@ -51,33 +33,6 @@ bool	parser(char *line)
 	// create_cmds()
 	// free_token_list()
 	return (false);
-}
-
-t_token	*get_last_token(t_token **head)
-{
-	t_token	*temp;
-
-	temp = *head;
-	while (temp->next)
-		temp = temp->next;
-	return (temp);
-}
-
-void	free_token_list(t_token **head)
-{
-	t_token	*curr;
-	t_token	*temp;
-
-	if (head == NULL)
-		return ;
-	curr = *head;
-	while (curr)
-	{
-		temp = curr->next;
-		free(curr->word);
-		free(curr);
-		curr = temp;
-	}
 }
 
 bool	validation_quart(char *line)
